+++ src/manage_report_exports.h
@@ -16,6 +16,88 @@
 #include <glib.h>
 
 /**
+ * @brief Size of the buffer used when copying generated report files.
+ */
+#define SCAN_REPORT_EXPORT_COPY_BUFFER_SIZE (64 * 1024)
+
+/**
+ * @brief Temporary and final files used by a scan report export.
+ */
+typedef struct
+{
+ /**
+  * Temporary working directory created for the export worker.
+  *
+  * Contains intermediate files generated during report processing and is
+  * removed after the export finishes or fails.
+  */
+ gchar *work_dir;
+
+ /**
+  * Path to the initial report XML generated by gvmd.
+  */
+ gchar *xml_start_path;
+
+ /**
+  * Path to the complete report XML used as input for report formatting.
+  */
+ gchar *xml_target_path;
+
+ /**
+  * Path to the formatted report file created in the temporary working
+  * directory.
+  */
+ gchar *formatted_path;
+
+ /**
+  * Final persistent path of the generated report export.
+  */
+ gchar *final_path;
+
+ /**
+  * MIME content type of the generated report file.
+  */
+ gchar *content_type;
+
+ /**
+  * File extension of the generated report file.
+  */
+ gchar *extension;
+
+ /**
+  * Size of the final report export file in bytes.
+  */
+ long long file_size;
+} report_export_files_t;
+
+/**
+ * @brief Temporary user context used by a report export worker.
+ */
+typedef struct
+{
+ /**
+  * UUID of the user active before switching to the export owner.
+  *
+  * Used to restore the previous user context after report generation.
+  */
+ gchar *previous_user_uuid;
+
+ /**
+  * UUID of the user who owns the report export.
+  *
+  * The worker switches to this user while generating the report.
+  */
+ gchar *export_user_uuid;
+
+ /**
+  * Whether the export user context is currently active.
+  *
+  * When TRUE, the previous user context must be restored during cleanup.
+  */
+ gboolean active;
+} report_export_user_context_t;
+
+/**
  * @brief Type of report export.
  */
 typedef enum
@@ -372,5 +454,35 @@ find_report_export_with_permission (const char *,report_export_t *,
 
 void
 manage_cleanup_old_report_exports (time_t);
+
+void
+report_export_files_cleanup (report_export_files_t *);
+
+void
+report_export_remove_final_file (const report_export_files_t *);
+
+int
+init_report_export_files (report_export_files_t *);
+
+int
+init_report_export_user_context (report_export_user_context_t *,
+                                 user_t);
+
+void
+cleanup_report_export_user_context (report_export_user_context_t *);
+
+void
+cleanup_report_export_get_data (get_data_t *);
+
+int
+format_report_export (const report_export_data_t,
+                      report_export_files_t *);
+
+int
+store_report_export_file (const report_export_data_t,
+                          report_export_files_t *);
+
+int
+check_report_export_cancel (report_export_t);
 
 #endif /* _GVMD_MANAGE_REPORT_EXPORTS_H */
