+++ src/manage_scan_report_exports.c
@@ -13,16 +13,12 @@
 #include "manage.h"
 #include "manage_report_configs.h"
 #include "manage_sql_report_exports.h"
-#include "manage_sql_report_formats.h"
 #include "manage_sql_users.h"
 
-#include <errno.h>
 #include <glib.h>
 #include <glib/gstdio.h>
 #include <stdio.h>
 #include <string.h>
-
-#include <gvm/util/fileutils.h>
 
 #undef G_LOG_DOMAIN
 
@@ -30,106 +26,6 @@
  * @brief GLib log domain.
  */
 #define G_LOG_DOMAIN "md manage"
-
-/**
- * @brief Size of the buffer used when copying generated report files.
- */
-#define SCAN_REPORT_EXPORT_COPY_BUFFER_SIZE (64 * 1024)
-
-/**
- * @brief Temporary and final files used by a scan report export.
- */
-typedef struct
-{
-  /**
-   * Temporary working directory created for the export worker.
-   *
-   * Contains intermediate files generated during report processing and is
-   * removed after the export finishes or fails.
-   */
-  gchar *work_dir;
-
-  /**
-   * Path to the initial report XML generated by gvmd.
-   */
-  gchar *xml_start_path;
-
-  /**
-   * Path to the complete report XML used as input for report formatting.
-   */
-  gchar *xml_target_path;
-
-  /**
-   * Path to the formatted report file created in the temporary working
-   * directory.
-   */
-  gchar *formatted_path;
-
-  /**
-   * Final persistent path of the generated report export.
-   */
-  gchar *final_path;
-
-  /**
-   * MIME content type of the generated report file.
-   */
-  gchar *content_type;
-
-  /**
-   * File extension of the generated report file.
-   */
-  gchar *extension;
-
-  /**
-   * Size of the final report export file in bytes.
-   */
-  long long file_size;
-} scan_report_export_files_t;
-
-/**
- * @brief Temporary user context used by a report export worker.
- */
-typedef struct
-{
-  /**
-   * UUID of the user active before switching to the export owner.
-   *
-   * Used to restore the previous user context after report generation.
-   */
-  gchar *previous_user_uuid;
-
-  /**
-   * UUID of the user who owns the report export.
-   *
-   * The worker switches to this user while generating the report.
-   */
-  gchar *export_user_uuid;
-
-  /**
-   * Whether the export user context is currently active.
-   *
-   * When TRUE, the previous user context must be restored during cleanup.
-   */
-  gboolean active;
-} scan_report_export_user_context_t;
-
-/**
- * @brief Get the directory for persistent report export files.
- *
- * @return Report export directory.
- */
-static const gchar *
-report_export_dir (void)
-{
-  static gchar *path = NULL;
-
-  if (path == NULL)
-    path = g_build_filename (GVMD_STATE_DIR,
-                             "report-exports",
-                             NULL);
-
-  return path;
-}
 
 /**
  * @brief Validate and resolve resources used by a scan report export.
@@ -330,152 +226,6 @@ manage_export_scan_report (const gchar *report_id,
 }
 
 /**
- * @brief Clean up scan report export file information.
- *
- * @param[in] files  File information to clean up.
- */
-static void
-scan_report_export_files_cleanup (scan_report_export_files_t *files)
-{
-  if (files == NULL)
-    return;
-
-  if (files->work_dir)
-    gvm_file_remove_recurse (files->work_dir);
-
-  g_free (files->work_dir);
-  g_free (files->xml_start_path);
-  g_free (files->xml_target_path);
-  g_free (files->formatted_path);
-  g_free (files->final_path);
-  g_free (files->content_type);
-  g_free (files->extension);
-
-  memset (files, 0, sizeof (*files));
-}
-
-/**
- * @brief Remove the completed file belonging to an export.
- *
- * @param[in] files  Report export file information.
- */
-static void
-scan_report_export_remove_final_file (
-  const scan_report_export_files_t *files)
-{
-  if (files && files->final_path)
-    g_unlink (files->final_path);
-}
-
-/**
- * @brief Initialize temporary paths used for report generation.
- *
- * @param[out] files  File information to initialize.
- *
- * @return 0 on success or -1 on failure.
- */
-static int
-init_scan_report_export_files (scan_report_export_files_t *files)
-{
-  GError *error;
-
-  if (files == NULL)
-    return -1;
-
-  memset (files, 0, sizeof (*files));
-
-  error = NULL;
-
-  files->work_dir = g_dir_make_tmp (
-    "gvmd-scan-report-export-XXXXXX",
-    &error);
-
-  if (files->work_dir == NULL)
-    {
-      g_warning ("%s: failed to create temporary directory: %s",
-                 __func__,
-                 error ? error->message : "Unknown error");
-
-      g_clear_error (&error);
-      return -1;
-    }
-
-  files->xml_start_path = g_build_filename (
-    files->work_dir,
-    "report-start.xml",
-    NULL);
-
-  files->xml_target_path = g_build_filename (
-    files->work_dir,
-    "report.xml",
-    NULL);
-
-  if (files->xml_start_path == NULL
-      || files->xml_target_path == NULL)
-    {
-      scan_report_export_files_cleanup (files);
-      return -1;
-    }
-
-  return 0;
-}
-
-/**
- * @brief Switch report generation to the export owner's user context.
- *
- * @param[out] context  User context to initialize.
- * @param[in]  owner    Owner of the report export.
- *
- * @return 0 on success or -1 on failure.
- */
-static int
-init_scan_report_export_user_context (
-  scan_report_export_user_context_t *context,
-  user_t owner)
-{
-  if (context == NULL || owner == 0)
-    return -1;
-
-  memset (context, 0, sizeof (*context));
-
-  context->previous_user_uuid = current_credentials.uuid;
-  context->export_user_uuid = user_uuid (owner);
-
-  if (context->export_user_uuid == NULL)
-    return -1;
-
-  current_credentials.uuid = context->export_user_uuid;
-  manage_session_init (current_credentials.uuid);
-
-  context->active = TRUE;
-
-  return 0;
-}
-
-/**
- * @brief Restore the user context active before report generation.
- *
- * @param[in] context  User context to restore.
- */
-static void
-cleanup_scan_report_export_user_context (
-  scan_report_export_user_context_t *context)
-{
-  if (context == NULL)
-    return;
-
-  if (context->active)
-    {
-      current_credentials.uuid = context->previous_user_uuid;
-      manage_session_init (current_credentials.uuid);
-    }
-
-  g_free (context->export_user_uuid);
-
-  memset (context, 0, sizeof (*context));
-}
-
-/**
  * @brief Initialize GET data used by the existing XML report generator.
  *
  * @param[in]  data  Stored report export data.
@@ -511,23 +261,6 @@ init_scan_report_export_get_data (
 }
 
 /**
- * @brief Clean up GET data initialized for report generation.
- *
- * @param[in] get  GET data to clean up.
- */
-static void
-cleanup_scan_report_export_get_data (get_data_t *get)
-{
-  if (get == NULL)
-    return;
-
-  g_free (get->type);
-  g_free (get->filter);
-
-  memset (get, 0, sizeof (*get));
-}
-
-/**
  * @brief Generate the intermediate XML used by a report format.
  *
  * @param[in] data      Report export data.
@@ -560,7 +293,7 @@ generate_scan_report_export_xml (
   if (init_scan_report_export_get_data (data, &get))
     return -1;
 
-  ret = manage_print_scan_report_xml_start (
+  ret = manage_print_report_xml_start (
     data->report,
     task,
     xml_path,
@@ -571,243 +304,9 @@ generate_scan_report_export_xml (
     data->ignore_pagination,
     data->lean);
 
-  cleanup_scan_report_export_get_data (&get);
+  cleanup_report_export_get_data (&get);
 
   return ret;
-}
-
-/**
- * @brief Apply the selected report format to generated scan report XML.
- *
- * @param[in]  data   Report export data.
- * @param[out] files  Report export file information.
- *
- * @return 0 on success or -1 on failure.
- */
-static int
-format_scan_report_export (
-  const report_export_data_t data,
-  scan_report_export_files_t *files)
-{
-  gchar *format_uuid;
-  GList *used_report_formats;
-
-  if (data == NULL
-      || files == NULL
-      || files->work_dir == NULL
-      || files->xml_start_path == NULL
-      || files->xml_target_path == NULL
-      || data->report_format == 0)
-    return -1;
-
-  if (report_format_active (data->report_format) == 0)
-    {
-      g_warning ("%s: report format is not active", __func__);
-      return -1;
-    }
-
-  if (report_format_predefined (data->report_format) == 0
-      && report_format_trust (data->report_format) != TRUST_YES)
-    {
-      g_warning ("%s: report format is not trusted", __func__);
-      return -1;
-    }
-
-  if (data->report_config
-      && report_config_report_format (data->report_config)
-      != data->report_format)
-    {
-      g_warning ("%s: report config is not compatible with report format",
-                 __func__);
-      return -1;
-    }
-
-  format_uuid = report_format_uuid (data->report_format);
-  if (format_uuid == NULL)
-    return -1;
-
-  used_report_formats = NULL;
-
-  files->formatted_path = apply_report_format (
-    format_uuid,
-    data->report_config,
-    files->xml_start_path,
-    files->xml_target_path,
-    files->work_dir,
-    &used_report_formats);
-
-  g_list_free (used_report_formats);
-  g_free (format_uuid);
-
-  if (files->formatted_path == NULL)
-    {
-      g_warning ("%s: report format returned no output file",
-                 __func__);
-      return -1;
-    }
-
-  return 0;
-}
-
-/**
- * @brief Build the final destination path of an exported report.
- *
- * @param[in] export_uuid  UUID of the report export.
- * @param[in] extension    File extension.
- *
- * @return Newly allocated path or NULL on failure.
- */
-static gchar *
-build_scan_report_export_path (const gchar *export_uuid,
-                               const gchar *extension)
-{
-  gchar *filename;
-  gchar *path;
-
-  if (str_blank (export_uuid) || extension == NULL)
-    return NULL;
-
-  if (g_mkdir_with_parents (report_export_dir (), 0700))
-    {
-      g_warning ("%s: failed to create %s: %s",
-                 __func__,
-                 report_export_dir (),
-                 strerror (errno));
-      return NULL;
-    }
-
-  if (extension[0])
-    filename = g_strdup_printf ("%s.%s",
-                                export_uuid,
-                                extension);
-  else
-    filename = g_strdup (export_uuid);
-
-  if (filename == NULL)
-    return NULL;
-
-  path = g_build_filename (report_export_dir (),
-                           filename,
-                           NULL);
-
-  g_free (filename);
-
-  return path;
-}
-
-/**
- * @brief Store a generated report in its persistent location.
- *
- * @param[in]  data   Report export data.
- * @param[out] files  Report export file information.
- *
- * @return 0 on success or -1 on failure.
- */
-static int
-store_scan_report_export_file (
-  const report_export_data_t data,
-  scan_report_export_files_t *files)
-{
-  GStatBuf stat_buffer;
-
-  if (data == NULL
-      || files == NULL
-      || data->uuid == NULL
-      || files->formatted_path == NULL)
-    return -1;
-
-  files->extension = report_format_extension (data->report_format);
-  files->content_type = report_format_content_type (data->report_format);
-
-  if (files->extension == NULL || files->content_type == NULL)
-    {
-      g_warning ("%s: report format metadata is incomplete",
-                 __func__);
-      return -1;
-    }
-
-  files->final_path = build_scan_report_export_path (
-    data->uuid,
-    files->extension);
-
-  if (files->final_path == NULL)
-    return -1;
-
-  /*
-   * Set the permissions before moving the file. The permissions are preserved
-   */
-  if (g_chmod (files->formatted_path, 0600))
-    {
-      g_warning ("%s: failed to set permissions on %s: %s",
-                 __func__,
-                 files->formatted_path,
-                 strerror (errno));
-      return -1;
-    }
-
-  if (gvm_file_move (files->formatted_path, files->final_path) == FALSE)
-    return -1;
-
-  if (g_stat (files->final_path, &stat_buffer))
-    {
-      g_warning ("%s: failed to stat %s: %s",
-                 __func__,
-                 files->final_path,
-                 strerror (errno));
-
-      g_unlink (files->final_path);
-      return -1;
-    }
-
-  files->file_size = (long long) stat_buffer.st_size;
-
-  return 0;
-}
-
-/**
- * @brief Check whether cancellation was requested.
- *
- * When cancellation is requested, the export is marked as canceled.
- *
- * @param[in] report_export  Report export.
- *
- * @return 0 to continue, 1 if canceled, or -1 on failure.
- */
-static int
-check_scan_report_export_cancel (report_export_t report_export)
-{
-  int ret;
-
-  ret = manage_report_export_cancel_requested (report_export);
-  if (ret < 0)
-    return -1;
-
-  if (ret == 0)
-    return 0;
-
-  if (manage_finish_report_export_cancel (report_export))
-    return -1;
-
-  return 1;
-}
-
-/**
- * @brief Mark a report export as failed.
- *
- * @param[in] report_export  Report export.
- * @param[in] message        Failure message.
- *
- * @return Always -1.
- */
-static int
-fail_scan_report_export (report_export_t report_export,
-                         const gchar *message)
-{
-  if (manage_fail_report_export (report_export, message))
-    g_warning ("%s: failed to mark report export as failed",
-             __func__);
-
-  return -1;
 }
 
 /**
@@ -821,8 +320,8 @@ int
 manage_process_scan_report_export (report_export_t report_export)
 {
   report_export_data_t data;
-  scan_report_export_files_t files;
-  scan_report_export_user_context_t user_context;
+  report_export_files_t files;
+  report_export_user_context_t user_context;
   gboolean final_file_stored;
   int ret;
 
@@ -842,7 +341,7 @@ manage_process_scan_report_export (report_export_t report_export)
 
   if (load_report_export_data (report_export, data))
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Failed to load report export data");
       goto cleanup;
@@ -851,7 +350,7 @@ manage_process_scan_report_export (report_export_t report_export)
   if (data->export_type != REPORT_EXPORT_TYPE_SCAN
       || data->delta_report != 0)
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Unsupported scan report export type");
       goto cleanup;
@@ -860,17 +359,17 @@ manage_process_scan_report_export (report_export_t report_export)
   if (data->status != REPORT_EXPORT_STATUS_RUNNING
       && data->status != REPORT_EXPORT_STATUS_CANCEL_REQUESTED)
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Report export is not running");
       goto cleanup;
     }
 
-  ret = check_scan_report_export_cancel (report_export);
+  ret = check_report_export_cancel (report_export);
 
   if (ret < 0)
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Failed to check report export cancellation");
 
@@ -881,30 +380,30 @@ manage_process_scan_report_export (report_export_t report_export)
   if (ret > 0)
     goto cleanup;
 
-  if (init_scan_report_export_user_context (&user_context,
-                                            data->owner))
+  if (init_report_export_user_context (&user_context,
+                                       data->owner))
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Failed to initialize report export user context");
       ret = -1;
       goto cleanup;
     }
 
-  if (init_scan_report_export_files (&files))
+  if (init_report_export_files (&files))
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Failed to initialize report export files");
       ret = -1;
       goto cleanup;
     }
 
-  ret = check_scan_report_export_cancel (report_export);
+  ret = check_report_export_cancel (report_export);
 
   if (ret < 0)
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Failed to check report export cancellation");
 
@@ -919,7 +418,7 @@ manage_process_scan_report_export (report_export_t report_export)
     report_export,
     REPORT_EXPORT_PROGRESS_GENERATING))
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Failed to update report export progress");
       ret = -1;
@@ -932,7 +431,7 @@ manage_process_scan_report_export (report_export_t report_export)
 
   if (ret)
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         ret == 2
           ? "Report filter was not found"
@@ -942,11 +441,11 @@ manage_process_scan_report_export (report_export_t report_export)
       goto cleanup;
     }
 
-  ret = check_scan_report_export_cancel (report_export);
+  ret = check_report_export_cancel (report_export);
 
   if (ret < 0)
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Failed to check report export cancellation");
 
@@ -957,9 +456,9 @@ manage_process_scan_report_export (report_export_t report_export)
   if (ret > 0)
     goto cleanup;
 
-  if (format_scan_report_export (data, &files))
+  if (format_report_export (data, &files))
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Failed to apply report format");
 
@@ -967,11 +466,11 @@ manage_process_scan_report_export (report_export_t report_export)
       goto cleanup;
     }
 
-  ret = check_scan_report_export_cancel (report_export);
+  ret = check_report_export_cancel (report_export);
 
   if (ret < 0)
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Failed to check report export cancellation");
 
@@ -982,9 +481,9 @@ manage_process_scan_report_export (report_export_t report_export)
   if (ret > 0)
     goto cleanup;
 
-  if (store_scan_report_export_file (data, &files))
+  if (store_report_export_file (data, &files))
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Failed to store generated report");
 
@@ -994,11 +493,11 @@ manage_process_scan_report_export (report_export_t report_export)
 
   final_file_stored = TRUE;
 
-  ret = check_scan_report_export_cancel (report_export);
+  ret = check_report_export_cancel (report_export);
 
   if (ret < 0)
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Failed to check report export cancellation");
 
@@ -1016,7 +515,7 @@ manage_process_scan_report_export (report_export_t report_export)
     files.content_type,
     files.extension))
     {
-      fail_scan_report_export (
+      manage_fail_report_export (
         report_export,
         "Failed to complete report export");
 
@@ -1029,10 +528,10 @@ manage_process_scan_report_export (report_export_t report_export)
 
 cleanup:
   if (ret != 0 && final_file_stored)
-    scan_report_export_remove_final_file (&files);
+    report_export_remove_final_file (&files);
 
-  scan_report_export_files_cleanup (&files);
-  cleanup_scan_report_export_user_context (&user_context);
+  report_export_files_cleanup (&files);
+  cleanup_report_export_user_context (&user_context);
   report_export_data_free (data);
 
   return ret;
